mark swig files as autogenerated
adds a .gitattributes file which should, according to GitHub[1],
cause the files autogenerated by Swig to be collapsed by
default in diffs as well as not contribute to the language
analysis.

I'm first testing out .gitattributes in this repo, and if it
works well, I'll replicate this commit to the other Medooze
repos.

[1]: https://docs.github.com/en/repositories/working-with-files/managing-files/customizing-how-changed-files-appear-on-github>